requirements.txt
Flask==1.0.2
google-cloud-pubsub
google-cloud-datastore
google-cloud-storage
oauth2client
google-api-python-client
google-cloud-tasks
google-cloud-bigquery
